pyproject.toml
[project]
name = "pillow-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pillow>=11.3.0",
    "pyside6>=6.9.2",
]
